+++ src/effects/spells/guidance.js
@@ -4,7 +4,7 @@ export function guidanceEffect(document) {
   let effect = baseSpellEffect(document, document.name);
   effect.changes.push(
     {
-      key: 'flags.midi-qol.optional.guidance.label.all',
+      key: 'flags.midi-qol.optional.guidance.label',
       mode: CONST.ACTIVE_EFFECT_MODES.CUSTOM,
       value: 'Guidance',
       priority: "20",
@@ -15,7 +15,14 @@ export function guidanceEffect(document) {
       value: '+ 1d4',
       priority: "20",
     },
+    {
+      key: 'flags.midi-qol.optional.guidance.skill.all',
+      mode: CONST.ACTIVE_EFFECT_MODES.CUSTOM,
+      value: '+ 1d4',
+      priority: "20",
+    },
   );
+  setProperty(effect, "flags.dae.specialDuration", ["isSkill", "isCheck"]);
 
   document.effects.push(effect);